+++ CHANGELOG.md
@@ -4,6 +4,7 @@ PowerModels.jl Change Log
 ### Staged
 - Added lambda-based convex hull relaxation scheme for trilinear products
 - Added QCWRTri Power Flow formulation
+- Added support for dual variables (KCL and branch thermal limits)
 
 ### v0.5.0
 - Standardized around branch name for pi-model lines (breaking)